# Switch op_now to be a json-op