Making it 20% faster and simpler.

### Benches

Comparing `performance.now()` (calling `op_now`) in the `deno_common` bench:

```
v1.8.2:
perf_now:                n = 500000, dt = 0.700s, r = 714286/s, t = 1400ns/op
Main:
perf_now:                n = 500000, dt = 0.403s, r = 1240695/s, t = 806ns/op
This PR:
perf_now:                n = 500000, dt = 0.331s, r = 1510574/s, t = 662ns/op
```